requirements.txt
pandas~=1.1.0
python-dotenv~=0.14.0
requests~=2.24.0
click~=7.1.2